Endpoint type refactor
